finished day 3 part 2
I used goroutines to handle parsing out the individual triangles. Was it performant? Not really. But it was a nice way to model the problem. A trio of goroutines read values from dedicated channels. They in turn parse the values, and generate triplets which get written to another channel. The parsing channel gets iterated over and writes a final value to a result channel, and the main goroutine waits for and returns the result.