day 4 part 1 complete
unit testing is super helpful when parsing is half your battle. Golang notoriously doesn't let you do repeated submatches, so the trick here is to do a first-order pass with a strings.Split() call, and run a regex on the final element to extract sector ID and checksum.

We form the checksum using a built-in sort with a supplied comparator.