day 8 part 1 complete
This one makes good use of a "space hash" -- a technique where I hash an image.Point instead of using a 2D array. The rect operation was pretty straightforward to code, particularly because the rect is always defined with the same origin as the screen.

For the rotate operation, I ended up using a double buffer, where I buffered the values from the screen for the given row/col, then overwrote that row/col with the shifted pixels, using a modulo operator to ensure pixels always wrapped around.