day 9 part 1 complete
this was a great chance to use a FSM to model the parser. Doing this let me count the nested markers as characters in compressed text and gave me better confidence that I'd hit all edge cases.

using the bytereader interface let me build code that was easy to test and adapted nicely to the actual file. Also, this is the first problem that hasn't used my library to get input, instead operating on the filesystem directly. neat!